+++ R/proc_chromR.R
@@ -47,13 +47,9 @@ proc.chromR <- function(x, win.size = 1e3, verbose=TRUE){
     warning( "annotation slot has no rows." )
   }
   
-  
-#  ptime <- system.time(x@seq.info$nuc.win <- regex.win(x))
   if(class(x@seq) == "DNAbin"){
     ptime <- system.time(x@seq.info$nuc.win <- seq2rects(x)) 
     if(verbose==TRUE){
-#      print("Nucleotide regions complete.")
-#      print(paste("  elapsed time: ", round(ptime[3], digits=4)))
       message("Nucleotide regions complete.")
       message(paste("  elapsed time: ", round(ptime[3], digits=4)))
     }
@@ -62,11 +58,8 @@ proc.chromR <- function(x, win.size = 1e3, verbose=TRUE){
   }
   
   if(class(x@seq) == "DNAbin"){
-#  ptime <- system.time(x@seq.info$N.win <- regex.win(x, regex="[n]"))
     ptime <- system.time(x@seq.info$N.win <- seq2rects(x, chars="n")) 
     if(verbose==TRUE){
-#      print("N regions complete.")
-#      print(paste("  elapsed time: ", round(ptime[3], digits=4)))
       message("N regions complete.")
       message(paste("  elapsed time: ", round(ptime[3], digits=4)))      
     }
@@ -76,15 +69,13 @@ proc.chromR <- function(x, win.size = 1e3, verbose=TRUE){
 
 
   # Population summary
-#  if(nrow(x@vcf.gt[x@var.info$mask,])>0){
-  if( nrow( x@vcf@gt[ x@var.info$mask, , drop = FALSE ] ) > 0 ){
-#    ptime <- system.time(x <- gt2popsum(x))
-    ptime <- system.time(x <- gt.to.popsum(x))
-    if(verbose==TRUE){
-#      print("Population summary complete.")
-#      print(paste("  elapsed time: ", round(ptime[3], digits=4)))
-      message("Population summary complete.")
-      message(paste("  elapsed time: ", round(ptime[3], digits=4)))
+  if( nrow(x@vcf@gt) > 0 ){
+    if( nrow( x@vcf@gt[ x@var.info$mask, , drop = FALSE ] ) > 0 ){
+      ptime <- system.time(x <- gt.to.popsum(x))
+      if(verbose==TRUE){
+        message("Population summary complete.")
+        message(paste("  elapsed time: ", round(ptime[3], digits=4)))
+      }
     }
   }
   
@@ -160,7 +151,7 @@ proc.chromR <- function(x, win.size = 1e3, verbose=TRUE){
 
   # Windowize variants.
 #  if(nrow(x@vcf.gt[x@var.info$mask,])>0){
-  if( nrow( x@vcf@gt[x@var.info$mask, , drop = FALSE ] ) > 0 ){
+  if( nrow( x@vcf@fix[x@var.info$mask, , drop = FALSE ] ) > 0 ){
     ptime <- system.time(x@win.info <- .windowize_variants(windows=x@win.info, variants=x@var.info[c('POS','mask')]))
     if(verbose==TRUE){
 #      print("windowize_variants complete.")