+++ tests/testthat/test_011_vcfR_methods.R
@@ -57,7 +57,8 @@ test_that("[samples = logical]",{
 test_that("show no gt",{
   data("vcfR_test")
   vcfR_test@gt <- matrix(nrow=0, ncol=0)
-#  head(vcfR_test)
+  myTest <-  capture.output(head(vcfR_test), type = c("output", "message"))
+  expect_equal(grep('No gt slot present', myTest), 23)
 })