+++ NEWS.md
@@ -10,6 +10,8 @@ This may have been addressed at 64a308ba50b9119108e8946737460de5997b805b by addi
 
 # vcfR 1.5.0.9000
 Released on CRAN 201X-XX-XX.
+* `rePOS()` attempts top create a non-overlapping coordinate system from POS and CHROM.
+* `vcfR2DNAbin()` manages the asterisk allele.
 * `extract.indels()` ignores GATK's <NON_REF>.
 * Added support for chromR objects with no gt slot to `proc.chromR()`.
 * Created `peak_to_ploid()` to call peaks and calculate dfe from `freq_peak()` output.